+++ src/FileExportModules/DOCX/InputsDocument/Boolean.ts
@@ -1,12 +1,15 @@
 import { Paragraph, TextRun } from 'docx';
 
+import { ConfigDTO } from '@Types/AppTypes';
 import { BooleanInputData } from '@Types/ProjectTypes';
 import { translations } from '@Translations/index';
-import ConfigService from '@Services/ConfigService';
 
-export function InputDocument_Boolean(inputData: BooleanInputData) {
+export function InputDocument_Boolean(o: {
+  config: ConfigDTO
+  inputData: BooleanInputData
+}) {
 
-  const R = translations.FileExportModules.docx[ConfigService.config.language];
+  const R = translations.FileExportModules.docx[o.config.language];
   const document: Paragraph[] = [];
 
   document.push(
@@ -18,7 +21,7 @@ export function InputDocument_Boolean(inputData: BooleanInputData) {
           color: '#000000',
           font: 'Calibri',
           size: `${12}pt`,
-          children: [ inputData.label ],
+          children: [ o.inputData.label ],
         }),
       ],
     })
@@ -32,7 +35,7 @@ export function InputDocument_Boolean(inputData: BooleanInputData) {
           font: 'Calibri',
           size: `${12}pt`,
           children: [
-            inputData.notApplicable === true ? R['Not applicable'] : String(inputData.value),
+            o.inputData.notApplicable === true ? R['Not applicable'] : String(o.inputData.value),
           ],
         }),
       ],