+++ src/FileExportModules/DOCX/InputsDocument/Selection.ts
@@ -1,12 +1,15 @@
 import { Paragraph, TextRun } from 'docx';
 
+import { ConfigDTO } from '@Types/AppTypes';
 import { SelectionInputData } from '@Types/ProjectTypes';
 import { translations } from '@Translations/index';
-import ConfigService from '@Services/ConfigService';
 
-export function InputDocument_Selection(inputData: SelectionInputData) {
+export function InputDocument_Selection(o: {
+  config: ConfigDTO
+  inputData: SelectionInputData
+}) {
 
-  const R = translations.FileExportModules.docx[ConfigService.config.language];
+  const R = translations.FileExportModules.docx[o.config.language];
   const document: Paragraph[] = [];
 
   document.push(
@@ -18,15 +21,15 @@ export function InputDocument_Selection(inputData: SelectionInputData) {
           color: '#000000',
           font: 'Calibri',
           size: `${12}pt`,
-          children: [ inputData.label ],
+          children: [ o.inputData.label ],
         }),
       ],
     })
   );
 
   let nothingSelected = true;
-  for (let i = 0; i < inputData.value.options.length; i++) {
-    if (inputData.value.options[i].id === inputData.value.id_selected) {
+  for (let i = 0; i < o.inputData.value.options.length; i++) {
+    if (o.inputData.value.options[i].id === o.inputData.value.id_selected) {
       nothingSelected = false;
       document.push(
         new Paragraph({
@@ -35,7 +38,7 @@ export function InputDocument_Selection(inputData: SelectionInputData) {
               color: '#000000',
               font: 'Calibri',
               size: `${12}pt`,
-              children: [inputData.value.options[i].optionLabel],
+              children: [o.inputData.value.options[i].optionLabel],
             }),
           ],
         })