+++ docs/ReleaseNotes.md
@@ -1,6 +1,5 @@
 // Before any commit, use the command line: npx tsc --noEmit
 
-# 1.9.32 translations ajustments
+# 1.9.33 Pictures info docx export
 
-- added translation for 'Download all pictures' button.
-- removed translations for feedback and services
+- Added formatting for picture input on docx exports